+++ packages/contracts/contracts/registry/AragonPluginRegistry.sol
@@ -4,14 +4,14 @@
 
 pragma solidity 0.8.10;
 
-import "../core/component/ERC165Registry.sol";
+import "../core/component/InterfaceBaseRegistry.sol";
 import "../core/IDAO.sol";
 import "../aragonPlugin/IPluginRepo.sol";
 
 /// @title Register plugin
 /// @author Sarkawt Noori - Aragon Association - 2022
 /// @notice This contract provides the possiblity to register a plugin pluginRepo by a unique address.
-contract AragonPluginRegistry is ERC165Registry {
+contract AragonPluginRegistry is InterfaceBaseRegistry {
     /// @notice Emitted if a new PluginRepo is registered
     /// @param name The name of the PluginRepo
     /// @param pluginRepo The address of the PluginRepo
@@ -21,7 +21,7 @@ contract AragonPluginRegistry is ERC165Registry {
     /// @param _dao the managing DAO address
     function initialize(IDAO _dao) public initializer {
         bytes4 pluginRepoInterfaceId = type(IPluginRepo).interfaceId;
-        __ERC165Registry_init(_dao, pluginRepoInterfaceId);
+        __InterfaceBaseRegistry_init(_dao, pluginRepoInterfaceId);
     }
 
     /// @notice Registers a PluginRepo