+++ packages/contracts/CHANGELOG.md
@@ -10,10 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added `IPluginFactory` abstract contract to be used in developing a pluginfactory.
 - added `IPluginRepo` interface for plugin PluginRepo contract.
 - added `PluginRepo` contract.
-- added `ERC165Registry` component to be used for creating any ERC165-based registry such as DAO-Registry and `AragonPluginRegistry`.
+- added `InterfaceBaseRegistry` component to be used for creating any ERC165-based registry such as DAO-Registry and `AragonPluginRegistry`.
 - added `AragonPluginRegistry` contract for registering `PluginRepo`s.
 - added `PluginRepoFactory` contract for creating `PluginRepo`s.
-- added testing for `PluginRepo`, `ERC165Registry`, `AragonPluginRegistry` and `PluginRepoFactory`.
+- added testing for `PluginRepo`, `InterfaceBaseRegistry`, `AragonPluginRegistry` and `PluginRepoFactory`.
 - added deployment script for `admin-dao`, `apm-registry` and `pluginRepo-factory`.
 
 ### Changed